+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/data/MovieRepositoryImpl.kt
@@ -1,14 +1,14 @@
 package edu.dyds.movies.data
 
 import edu.dyds.movies.data.external.MovieDetailExternalSource
-import edu.dyds.movies.data.external.tmdb.TMDBMoviesExternalSource
+import edu.dyds.movies.data.external.PopularMoviesExternalSource
 import edu.dyds.movies.data.local.MoviesLocalSource
 import edu.dyds.movies.domain.entity.Movie
 import edu.dyds.movies.domain.repository.MovieRepository
 
 class MovieRepositoryImpl(
     private val moviesLocalSource: MoviesLocalSource,
-    private val popularMoviesExternalSource: TMDBMoviesExternalSource,
+    private val popularMoviesExternalSource: PopularMoviesExternalSource,
     private val movieDetailExternalSource: MovieDetailExternalSource
 ) : MovieRepository {
