+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/data/external/tmdb/TMDBMoviesExternalSourceImpl.kt
@@ -1,6 +1,7 @@
 package edu.dyds.movies.data.external.tmdb
 
 import edu.dyds.movies.data.external.MovieDetailExternalSource
+import edu.dyds.movies.data.external.PopularMoviesExternalSource
 import edu.dyds.movies.domain.entity.Movie
 import io.ktor.client.*
 import io.ktor.client.call.*
@@ -11,7 +12,7 @@ import kotlinx.serialization.Serializable
 
 class TMDBExternalSourceImpl(
     private val tmdbHttpClient: HttpClient,
-) : MovieDetailExternalSource, TMDBMoviesExternalSource {
+) : MovieDetailExternalSource, PopularMoviesExternalSource {
 
     override suspend fun getPopularMovies(): List<Movie> {
         val remoteResult: RemoteResult = tmdbHttpClient.get("/3/discover/movie?sort_by=popularity.desc").body()
